% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bma.df1.R
\name{bma.df1}
\alias{bma.df1}
\title{Re-name data and call BMA function for 1-degree-of-freedom Test}
\usage{
bma.df1(Sample = NULL, Covar = NULL, cc = 0.5, co = 0.5, phi = 1,
  psi = 1000, formul = NULL)
}
\arguments{
\item{Sample}{input A data frame or matrix with columns in the order: "Y" "E" "G". Defaults to NULL.}

\item{Covar}{input A data frame of all covariate variables. Defaults to NULL.}

\item{cc}{input A number. Defaults to 0.5.}

\item{co}{input A number. Defaults to 0.5.}

\item{phi}{input A number. Defaults to 1.}

\item{psi}{input A number. Defaults to 1000.}

\item{formul}{input A string in the form: "Count~E+G+E:G+Y+E:Y+G:Y+E:G:Y+... . Defaults to NULL.}
}
\value{
Vector containing BMA interaction effect estimate, standard error, Z-score, and p-value.
}
\description{
Re-name data and call BMA function for 1-degree-of-freedom Test
}
\examples{
data("Covariates")
data("Sample")
formul <- "Count~E+G+E:G+Y+E:Y+G:Y+E:G:Y+C1+C2+C3+Y:C1+Y:C2+Y:C3+E:C2"
bma.df1(Sample=Sample,Covar=Covariates,cc=0.5,co=0.5,phi=1,psi=1000,formul=NULL)
bma.df2(Sample=Sample,Covar=Covariates,cc=0.5,co=0.5,phi=1,psi=1000,formul=NULL)
bma.df1(Sample=Sample,Covar=Covariates,cc=0.5,co=0.5,phi=1,psi=1000,formul=formul)
bma.df2(Sample=Sample,Covar=Covariates,cc=0.5,co=0.5,phi=1,psi=1000,formul=formul)
}
