% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Covariates}
\alias{Covariates}
\title{Covariates}
\format{A data frame with 1000 rows and 3 variables:
\describe{
  \item{Covariate 1}{Covariate,binary}
  \item{Covariate 2}{Covariate, binary}
  \item{Covariate 3}{Covariate, binary}
  ...
}}
\usage{
Covariates
}
\description{
A dataset containing Covariate 1,Covariate 2, and Covariate 3 on 1000 individuals
}
\keyword{datasets}
