% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.BMA.1DF.R
\name{run.BMA.1DF}
\alias{run.BMA.1DF}
\title{BMA function for 1-degree-of-freedom Test}
\usage{
run.BMA.1DF(Y = NULL, E = NULL, G = NULL, Cov = NULL, cc = 0.05,
  co = 0.05, phi = 1, psi = 1000, formul = NULL)
}
\arguments{
\item{Y}{input A vector of binary outcomes/phenotypes. Defaults to NULL.}

\item{E}{input A vector of binary exposure status. Defaults to NULL.}

\item{G}{input A vector of binary genotype. Defaults to NULL.}

\item{Cov}{of all covariate variables. Defaults to NULL.}

\item{cc}{input A number. Defaults to 0.5.}

\item{co}{input A number. Defaults to 0.5.}

\item{phi}{input A number. Defaults to 1.}

\item{psi}{input A number. Defaults to 1000.}

\item{formul}{input A string in the form: "Count~E+G+E:G+Y+E:Y+G:Y+E:G:Y+... . Defaults to NULL.}
}
\value{
Vector containing BMA interaction effect estimate, standard error, Z-score, and p-value.
}
\description{
BMA function for 1-degree-of-freedom Test
}
