% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sample}
\alias{Sample}
\title{Phenotype, Environmental Exposure, and G in dominant form}
\format{A data frame with 1000 rows and 3 variables:
\describe{
  \item{Y}{Phenotype,binary}
  \item{E}{Environmental factor, binary}
  \item{G}{Genetic factor, binary}
  ...
}}
\usage{
Sample
}
\description{
A dataset containing Y,E, and G on 1000 individuals
}
\keyword{datasets}
