% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.R
\name{wald}
\alias{wald}
\title{The P-value from a Wald Test}
\usage{
wald(est, se)
}
\arguments{
\item{est}{input A number.}

\item{se}{input A number.}
}
\value{
The P-value.
}
\description{
The P-value from a Wald Test
}

